test discounts

+++ application/modules/mod_discount/models/discount_model_admin.php
@@ -35,7 +35,12 @@ class Discount_model_admin extends CI_Model {
         if ( $active == 1)
             $active = 0;
         else $active=1;
-
+        
+        // Check is discount with such id
+        if ($discount == null)
+            return false;
+        
+        // If updated active succes then return TRUE
         if ($this->db->where('id',$id)->update('mod_shop_discounts',array('active'=>$active)))
             return true;
         
@@ -74,6 +79,7 @@ class Discount_model_admin extends CI_Model {
      * return boolean|array
      */
     public function getUsersByIdNameEmail($term, $limit = 7) {
+        
         $query = $this->db
                 ->like('username', $term)
                 ->or_like('email', $term)
@@ -142,12 +148,14 @@ class Discount_model_admin extends CI_Model {
      * @return boolean|int
      */
     public function insertDataToDB($tableName , $data) {
-        try {
+        
             $this->db->insert($tableName, $data);
-            return $this->db->insert_id();
-        }catch (Exception $e) {
-            return false;
-        }
+            $id =  $this->db->insert_id();
+            if ($id)
+                return $id;
+            else
+                return false;
+        
     }
     
     /**
@@ -189,7 +197,6 @@ class Discount_model_admin extends CI_Model {
             $query = $query->where('discount_id !=',$editDiscountId);
         
         $query = $query->where('end_value',null)->or_where('end_value',0)->get('mod_discount_comulativ')->result_array();
-        
         if (count($query))
             return true;
         else 
@@ -263,8 +270,9 @@ class Discount_model_admin extends CI_Model {
     public function deleteDiscountById($id) {
         $query = $this->db->from('mod_shop_discounts')->where('id',$id)->get()->row_array();
         $discountType = $query['type_discount'];
-        
-         try {
+        if (!$query)
+            return false;
+        try {
             $this->db->where('id',$id)->delete('mod_shop_discounts');
             $this->db->where('discount_id',$id)->delete('mod_discount_'.$discountType);
             return true;
